Fix type check with latest mypy/typeshed
The following commit defined a return type hint for
getElementsByTagName:

https://github.com/python/typeshed/commit/3fc2f27990e2a9de0c86f16c2bdf5f80d84720a4#diff-f451f731d037ef9d79347194490b32ba613798ea7eaa2c160351a69625f05e08R150

It defined the return type as a list of Node, while this code expects a
list of Element (Element is a subtype of Node).

Given that one would expect a getElements method to return
specifically *elements* and not other types of node, I think the
typeshed change may be incorrect, but it's hard to be sure since the
stdlib docs themselves are ambiguous.

Suppress it for now to unblock dependency updates.