Ensure directories appear first in listings [RHELDST-21890]
Directories are generally expected to be listed first in directory
indexes. That was already working for yum and file repos, but wasn't the
case for kickstart repos due to their combination of different types of
content.

This commit applies a consistent sorting so that directories will always
come first, and entries will otherwise be sorted by name, for all repo
types.